Ensure Webservice ConfigMap names references align
Uses the same template to calculate:

1. The name of the Webservice ConfigMap.
2. The reference to the name of the Webservice ConfigMap in the
   Webservice Deployment.

Normally this misalignment wasn't a problem, but if the release name
contains the word "webservice", then the "webservice.fullname" template
used in the calculation of the ".fullname" variable (formed in the
"webservice.datamodel.prepare" template) would trim out the redundant
"webservice" string. This meant the Deployment would be looking for the
ConfigMap by the wrong name.

First seen in
https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/3089#note_1352012712.

Changelog: fixed